move serializer and contribution to common/, delete old browser/ and parser modules
- Rewrite quartoNotebookSerializer.ts as QuartoNotebookSerializer class
  using the new notebookToQmd/qmdToNotebook modules
- Move quartoNotebookContribution.ts from browser/ to common/
- Delete old quartoNotebookParser.ts (replaced by qmdToNotebook.ts)
- Update workbench.common.main.ts import path
